size doesn't matter

+++ mch2022/components/mch2022_rp2040/mch2022_rp2040.cpp
@@ -31,14 +31,14 @@ void Mch2022_rp2040Component::dump_config() {
 }
 
 #ifdef USE_BINARY_SENSOR
-void Mch2022_rp2040Component::set_sub_binary_sensor(uint8_t input, binary_sensor::BinarySensor *sens) {
+void Mch2022_rp2040Component::set_sub_binary_sensor(size_t input, binary_sensor::BinarySensor *sens) {
 // void Mch2022_rp2040Component::set_sub_binary_sensor(uint8_t input, binary_sensor::BinarySensor *sens) {
   // if (input < SubBinarySensorInput::SUB_BINARY_SENSOR_INPUT_COUNT) {
-  this->sub_binary_sensors_[input] = sens;
+  this->sub_binary_sensors_[(size_t)input] = sens;
   //}
 }
 
-void Mch2022_rp2040Component::update_sub_binary_sensor_(uint8_t input, bool value) {
+void Mch2022_rp2040Component::update_sub_binary_sensor_(size_t input, bool value) {
   // size_t index = (size_t) input;
   if (
     (this->sub_binary_sensors_[input] != nullptr) &&
